多个文件

+++ src/main/java/com/qiuhua/kkk/dto/questionDTO.java
@@ -1,19 +1,20 @@
 package com.qiuhua.kkk.dto;
 
-import com.qiuhua.kkk.Model.User;
+import com.qiuhua.kkk.model.User;
 import lombok.Data;
 @Data
 public class questionDTO {
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
-    private Integer creatId;
+    private Long creatId;
     private Long gmtCreat;
     private Long gmtModified;
     private Integer viewCount;
     private Integer likeCount;
     private Integer commentCount;
     private  String tag;
+
     private  User user;
 
 }
